Support Audiothingies P6 port allocation for I2C
In order to minimize changes, we let MIOS32_IIC_NUM (number of I2C
ports) be 2 as for STM32F4DISCOVERY, but we don't initialize the first port
(#0, mapped to I2C2) as the corresponding pins are used for other things
in the P6. The EEPROM in the P6 is connected to I2C1 which thus is
device #1 (of #0 and #1).