Use P6 on-board EEPROM (24LC512) to store and load patches
This is done using the MIOS32 banksticks interface, which is really
intended for removable 24LC256/24LC512 chips, but we can use the same
interface for the on-board EEPROM, which incidentally is a 24LC512 anyway.

The on-board P6 24LC512 comes across as a single chip, using IIC port 1,
with 1 bank (i.e. 1 chip) available.